fixed variable position spawning in main
